RHS  (#225)
* updated right sider

* updated with new changes on rightsider

* prop added

* updated RHS side with functions

* removed unwanted variables

* updated group loading

* vue warn fixed

* updated
