+++ camunda-formio-tasklist-vue/src/components/TaskList.vue
@@ -39,10 +39,13 @@
             :filterList="filterList"
           />
         </div>
+       <!-- need to bring here right side -->
         <div
           class="ctf-task-details-container ms-md-2 rounded"
           :class="{ 'col-12 mx-0': !maximize ,'col-md-9 col-12':maximize}"
         >
+
+       <!-- single taks loading -->
           <div
             v-if="singleTaskLoading"
             class="d-flex justify-content-center align-items-center"
@@ -54,430 +57,30 @@
               <span class="sr-only">Loading...</span>
             </div>
           </div>
-          <template v-else-if="getFormsFlowTaskId && task">
-            <ExpandContract />
-            <div
-              class="bg-primary task-title"
-              ref="taskTitleRef"
-            >
-              <h3
-                class="m-0"
-                data-bs-toggle="tooltip"
-                title="Task Name"
-              >{{ task.name }}</h3>
-            </div>
-            <div
-              class="d-flex flex-column w-100 px-4 py-2 task-details"
-              :style="{
-                height: taskScrollableHeight
-              }"
-            >
-              <div class="d-flex mb-1">
-                <h4
-                  class="mt-2 mb-3"
-                  data-bs-toggle="tooltip"
-                  title="Process Name"
-                >{{ task.taskProcess }}
-                </h4>
-                <p
-                  class="mt-2 mb-3"
-                  v-if="task.applicationId"
-                  data-bs-toggle="tooltip"
-                  title="Application Id"
-                > <span class="mx-4">|</span>Application ID <strong>#{{ task.applicationId }}</strong>
-                </p>
-              </div>
-              <div class="d-flex justify-content-start mb-4">
-                <section v-if="!hideTaskDetails.assignee" class="task-assignee">
-                  <label class="fw-bold">Task assignee</label>
-                  <button
-                    v-if="task.assignee"
-                    class="btn task-icon-btn"
-                    @click="toggleassignee"
-                    data-bs-toggle="tooltip"
-                    title="Click to change assignee"
-                  >
-                   <span v-if="loadingEditAssignee" class="spinner-grow spinner-grow-sm" role="status" aria-hidden="true"></span>
-                    <i v-if="!loadingEditAssignee"
-                      class="fa fa-pencil"
-                      :class="{
-                      'fa-times-circle-o':editAssignee,
-                      'fa-pencil': !editAssignee
-                    }"
-                    />
-                  </button>
-                  <div class="d-flex align-items-baseline">
-                    <template v-if="task.assignee">
-                      <div
-                        v-if="editAssignee"
-                        class="d-flex w-100 mt-1"
-                      >
-                        <v-select
-                          :label="selectSearchType!=='email'?selectSearchType:'code'"
-                          :options="reviewerUsersList"
-                          :filterable="false"
-                          v-model="userSelected"
-                          :placeholder="`Search by ${selectSearchType}`"
-                          @search="onUserSearch"
-                          class="select-assignee"
-                        >
-                         <template v-if="selectSearchType==='email' " v-slot:option="option">
-                          <div>{{option.code}}</div>
-                          <div>({{option.email}})</div>
-                         </template>
-                          </v-select>
-                        <div class="dropdown assignee-search-filter mx-2">
-                          <button
-                            class="btn btn-secondary dropdown-toggle"
-                            type="button"
-                            data-bs-toggle="dropdown"
-                            aria-expanded="false"
-                          >
-                            <i class="fa fa-filter" />
-                          </button>
-                          <ul
-                            class="dropdown-menu"
-                            aria-labelledby="dropdownMenuButton1"
-                          >
-                            <li
-                              v-for="(row, index) in UserSearchListLabel"
-                              :key="row.id"
-                              @click="setSelectedUserSearchBy(row.searchType, index)"
-                              class="dropdown-item"
-                              :class="{'active': index === activeUserSearchindex}"
-                            >
-                              {{ row.label }}
-                            </li>
-                          </ul>
-                        </div>
-                        <button
-                          class="btn task-icon-btn btn-light"
-                          @click="onSetassignee"
-                          data-bs-toggle="tooltip"
-                          title="Set assignee"
-                        >
-                          <i class="fa fa-check-circle-o fa-lg" />
-                        </button>
-                      </div>
-                      <template v-else>
-                        <div class="assignee-name">{{ task.assignee }}</div>
-                        <button
-                          class="btn task-icon-btn"
-                          @click="onUnClaim"
-                          data-bs-toggle="tooltip"
-                          title="Reset assignee"
-                        > 
-                         <span v-if="loadingClaimAndUnclaim" class="spinner-grow spinner-grow-sm" role="status" aria-hidden="true"></span>
+       <!-- single task loading end -->
 
-                          <i v-if="!loadingClaimAndUnclaim" class="fa fa-times-circle-o" />
-                        </button>
-                      </template>
-                    </template>
-                    <button
-                      v-else
-                      class="btn btn-light"
-                      @click="onClaim"
-                      data-bs-toggle="tooltip"
-                      title="Claim task"
-                    >
-                     <span v-if="loadingClaimAndUnclaim" class="spinner-grow spinner-grow-sm" role="status" aria-hidden="true"></span>
-                     <i v-if="!loadingClaimAndUnclaim" class="fa fa-plus" />
-                      <span class="mx-1">Claim</span>
-                    </button>
-                  </div>
-                </section>
-                <section
-                v-if="!hideTaskDetails.group" 
-                  class="task-groups mx-4"
-                  data-bs-toggle="tooltip"
-                  title="Click to modify groups"
-                >
-                  <label class="fw-bold">Groups</label>
-                  <button
-                    v-if="groupListNames && groupListNames.length"
-                    class="btn task-icon-btn"
-                    data-bs-toggle="modal"
-                    data-bs-target="#groupsModal"
-                  >
-                    <i class="fa fa-pencil" />
-                  </button>
-                  <div class="d-flex align-items-baseline group-name">
-                    <template v-if="groupListNames && groupListNames.length">
-                     <p class="text-truncate"> {{ String(groupListNames) }}</p>
-                    </template>
-                    <button
-                      v-else
-                      class="btn btn-light"
-                      data-bs-toggle="modal"
-                      data-bs-target="#groupsModal"
-                    >
-                      <i class="fa fa-plus" />
-                      <span class="mx-1">Add Groups</span>
-                    </button>
-                  </div>
-                  <div
-                    class="modal fade task-groups-modal"
-                    id="groupsModal"
-                    tabindex="-1"
-                    aria-labelledby="groupsModal"
-                    aria-hidden="true"
-                  >
-                    <div class="modal-dialog">
-                      <div class="modal-content">
-                        <div class="modal-header">
-                          <h5 class="modal-title">MANAGE GROUPS</h5>
-                          <i
-                            type="button"
-                            class="fa fa-times mx-2"
-                            data-bs-dismiss="modal"
-                            aria-label="Close"
-                            title="cancel"
-                          ></i>
-                        </div>
-                        <div class="modal-body px-4 pb-5">
-                          <i class="fa fa-exclamation-circle"></i>
-                          You can add a group by typing a group ID into the input
-                          field and afterwards clicking the button with the plus sign.
-                          <div class="d-flex my-3">
-                            <input
-                              class="form-control group-name-input"
-                              type="text"
-                              v-model="setGroup"
-                              @keyup.enter="addGroup"
-                              placeholder="Group ID"
-                            />
-                            <button
-                              class="btn btn-primary add-group-btn"
-                              @click="addGroup"
-                              :disabled="!setGroup"
-                            >
-                              <i class="fa fa-plus" />
-                              <span class="mx-1">Add group</span>
-                            </button>
-                          </div>
-                          <div
-                            v-if="groupList.length"
-                            class="d-flex flex-wrap"
-                          >
-                            <div
-                              class="d-flex align-items-baseline added-group-chip"
-                              v-for="g in groupList"
-                              :key="g.groupId"
-                              @click="deleteGroup(g.groupId)"
-                              data-bs-toggle="tooltip"
-                              title="Click to remove this group"
-                            >
-                              <div class="mx-1">{{ g.groupId }}</div>
-                              <i class="fa fa-times-circle-o" />
-                            </div>
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                </section>
-              </div>
-              <div class="d-flex justify-content-between mb-2">
-                <section class="task-date-picker" v-if="!hideTaskDetails.followUpDate" >
-                  <label class="fw-bold mb-1">Follow up</label>
-                  <div
-                    class="d-flex align-items-baseline"
-                    v-if="task.followUp"
-                  >
-                    <p
-                      data-bs-toggle="tooltip"
-                      :title="getExactDate(task.followUp)"
-                    >{{ timedifference(task.followUp) }}</p>
-                    <button
-                      class="btn task-icon-btn"
-                      @click="removeFollowupDate"
-                      data-bs-toggle="tooltip"
-                      title="Click to remove FollowUp Date"
-                    >
-                      <i class="fa fa-times-circle-o" />
-                    </button>
-                  </div>
-                  <v-date-picker
-                    v-else
-                    v-model="task.followUp"
-                    :popover="{ visibility: 'click' }"
-                    @input="updateFollowUpDate"
-                  >
-                    <template v-slot="{ inputValue, inputEvents }">
-                      <div class="input-group">
-                        <input
-                          class="form-control"
-                          :value="inputValue"
-                          v-on="inputEvents"
-                          placeholder="mm/dd/yyyy"
-                        />
-                        <i class="fa fa-calendar-alt"></i>
-                      </div>
-                    </template>
-                  </v-date-picker>
-                </section>
-                <section class="task-date-picker" v-if="!hideTaskDetails.dueDate" >
-                  <label class="fw-bold mb-1">Due date</label>
-                  <div
-                    class="d-flex align-items-baseline"
-                    v-if="task.due"
-                  >
-                    <p
-                      data-bs-toggle="tooltip"
-                      :title="getExactDate(task.due)"
-                    >{{ timedifference(task.due) }}</p>
-                    <button
-                      class="btn task-icon-btn"
-                      @click="removeDueDate"
-                      data-bs-toggle="tooltip"
-                      title="Click to remove Due date"
-                    >
-                      <i class="fa fa-times-circle-o" />
-                    </button>
-                  </div>
-                  <v-date-picker
-                    v-else
-                    v-model="task.due"
-                    :popover="{ visibility: 'click' }"
-                     @input="updateDueDate"
-                  >
-                    <template v-slot="{ inputValue, inputEvents }">
-                      <div class="input-group">
-                        <input
-                          class="form-control"
-                          :value="inputValue"
-                          v-on="inputEvents"
-                          placeholder="mm/dd/yyyy"
-                        />
-                        <i class="fa fa-calendar-alt"></i>
-                      </div>
-                    </template>
-                  </v-date-picker>
-                </section>
-                <section v-if="!hideTaskDetails.createdDate" >
-                  <label class="fw-bold mb-1">Created</label>
-                  <p
-                    data-bs-toggle="tooltip"
-                    :title="getExactDate(task.created)"
-                  >{{ timedifference(task.created) }}</p>
-                </section>
-                <section></section>
-              </div>
-              <ul
-                class="nav nav-tabs mt-3 task-tabs"
-                role="tablist"
-              >
-                <li
-                  class="nav-item"
-                  role="presentation"
-                >
-                  <button
-                    class="nav-link active"
-                    id="task-form-tab"
-                    data-bs-toggle="tab"
-                    data-bs-target="#taskForm"
-                    type="button"
-                    role="tab"
-                    aria-controls="form"
-                    aria-selected="true"
-                  >Form</button>
-                </li>
-                <li
-                  class="nav-item"
-                  role="presentation"
-                >
-                  <button
-                    class="nav-link"
-                    id="task-history-tab"
-                    data-bs-toggle="tab"
-                    data-bs-target="#taskHistory"
-                    type="button"
-                    role="tab"
-                    aria-controls="history"
-                    aria-selected="false"
-                  >History</button>
-                </li>
-                <li
-                  class="nav-item"
-                  role="presentation"
-                >
-                  <button
-                    class="nav-link"
-                    id="task-diagram-tab"
-                    data-bs-toggle="tab"
-                    data-bs-target="#diagramContainer"
-                    type="button"
-                    role="tab"
-                    aria-controls="diagram"
-                    aria-selected="false"
-                    @click="getDiagramDetails"
-                  >Diagram</button>
-                </li>
-              </ul>
-              <div class="tab-content py-3 task-tab-content">
-                <!-- Form tab content -->
-                <div
-                  class="tab-pane fade show active form-tab-content"
-                  id="taskForm"
-                  role="tabpanel"
-                  aria-labelledby="form-tab"
-                  :class="{'disabled': task.assignee !== userName}"
-                >
-                  <div v-if="task.assignee === userName">
-                    <transition name="fade">
-                      <FormEdit
-                        :formioUrl="formioUrl"
-                        @onformsubmit="onFormSubmitCallback"
-                        @oncustomevent="oncustomEventCallback"
-                      />
-                    </transition>
-                  </div>
-                  <div v-else>
-                    <FormView :formioUrl="formioUrl" />
-                  </div>
-                </div>
-                <!-- Form tab content end -->
-                <div
-                  class="tab-pane fade"
-                  id="taskHistory"
-                  role="tabpanel"
-                  aria-labelledby="history-tab"
-                >
-                  <TaskHistory
-                    :taskHistoryList="taskHistoryList"
-                    :applicationId="task.applicationId"
-                  />
-                </div>
-                <div
-                  class="tab-pane fade"
-                  id="diagramContainer"
-                  role="tabpanel"
-                  aria-labelledby="diagram-tab"
-                >
-                  <div
-                    v-if="diagramLoading"
-                    class="d-flex justify-content-center"
-                  >
-                    <div
-                      class="spinner-border"
-                      role="status"
-                    >
-                      <span class="sr-only">Loading...</span>
-                    </div>
-                  </div>
-                    <div
-                    v-show="!diagramLoading"
-                    class="cft-bpmn-viewer-container"
-                  >
-                    <div
-                      class="cft-bpm-container cft-grab-cursor"
-                      id="process-diagram-container"
-                    ></div>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </template>
+      <!-- right side if not loading  -->
+    <RightSider
+  v-else-if="getFormsFlowTaskId && task"
+  :task="task"
+  :bpmApiUrl="bpmApiUrl"
+  :token="token"
+  :getBPMTaskandReload="getBPMTaskandReload"
+  :taskScrollableHeight="taskScrollableHeight"
+  :reloadCurrentTask="reloadCurrentTask"
+  :getDiagramDetails="getDiagramDetails"
+  :formioUrl="formioUrl"
+  :onFormSubmitCallback="onFormSubmitCallback"
+  :oncustomEventCallback="oncustomEventCallback"
+  :taskHistoryList="taskHistoryList"
+  :diagramLoading="diagramLoading"
+  :userName="userName"
+  :hideTaskDetails="hideTaskDetails"
+/>
+
+      <!-- right side end  -->
+
+      <!-- task not selected start -->
           <div
             v-else
             class="d-flex align-items-center justify-content-center task-details-empty"
@@ -488,7 +91,12 @@
             <i class="fa fa-exclamation-circle"></i>
             <h4 class="mt-0 mx-2">Please select a task from the list</h4>
           </div>
+      <!-- task not selected end -->
+
+
         </div>
+       <!-- end  right side -->
+
       </div>
     </template>
     <div v-else>
@@ -530,18 +138,14 @@ import * as bootstrap from 'bootstrap';
 import {
   ALL_FILTER,
   CamundaRest,
-  SEARCH_USERS_BY,
   SocketIOService,
   authenticateFormio, 
   findFilterIdForDefaultFilterName,
   getFormDetails,
-  getFormattedDateAndTime,
-  getISODateTime,
   getTaskFromList,
   getUserName,
   getformHistoryApi,
   isAllowedUser,
-  reviewerGroup,
   sortByPriorityList,
 } from "../services";
 import {
@@ -552,53 +156,41 @@ import {
   FilterPayload,
   FormRequestActionPayload,
   FormRequestPayload,
-  GroupListPayload,
   Payload,
-  SEARCH_OPTION_TYPE,
   TaskHistoryListPayload,
   TaskListSortType,
   TaskPayload,
-  UserListObject,
-  UserListPayload,
-  UserPayload,
-  UserSearchListLabelPayload,
 } from "../models";
 import  {
   Toast,
   Tooltip
 } from 'bootstrap';
 import BpmnViewer from "bpmn-js/dist/bpmn-navigated-viewer.production.min.js";
-import DatePicker from "v-calendar/lib/components/date-picker.umd";
 import ExpandContract from "./addons/ExpandContract.vue";
-import FormEdit from "./form/Edit.vue";
-import FormView from "./form/View.vue";
+
 import {
   Formio
 } from 'vue-formio';
 import Header from "./layout/Header.vue";
 import LeftSider from "./layout/LeftSider.vue";
-import TaskHistory from "../components/addons/TaskHistory.vue";
+import RightSider from "../components/layout/RightSider.vue";
 import TaskListMixin from "../mixins/TaskListMixin.vue";
-import moment from "moment";
+
 import {
   namespace
 } from "vuex-class";
 import serviceFlowModule from "../store/modules/serviceFlow-module";
-import vSelect from "vue-select";
 
 const StoreServiceFlowModule = namespace("serviceFlowModule");
 
 @Component({
   components: {
-    TaskHistory,
     Header,
     LeftSider,
-    vSelect,
     ExpandContract,
     BpmnViewer,
-    FormEdit,
-    FormView,
-    VDatePicker: DatePicker,
+    RightSider,
+ 
   },
 })
 export default class Tasklist extends Mixins(TaskListMixin) {
@@ -642,17 +234,9 @@ export default class Tasklist extends Mixins(TaskListMixin) {
   private formioUrl: string = "";
   private task: TaskPayload = {
   };
-  private setGroup = null;
-  private userSelected: UserListPayload = {
-  };
+
   public perPage: number = 10;
   private filterList: FilterPayload[] = [];
-  private editAssignee: boolean = false;
-  private loadingEditAssignee: boolean=  false;
-  private loadingClaimAndUnclaim: boolean = false;
-  private groupList: GroupListPayload[] = [];
-  private groupListNames?: string[] = [];
-  private groupListItems: string[] = [];
   private userEmail: string = "external";
   private selectedfilterId: string = "";
   private xmlData!: string;
@@ -662,12 +246,11 @@ export default class Tasklist extends Mixins(TaskListMixin) {
     maxResults: this.perPage,
   };
   private taskHistoryList: TaskHistoryListPayload[] = [];
-  private reviewerUsersList: UserListPayload[] = [];
-  private selectSearchType: string = "lastName";
+
   private taskIdValue: string = "";
   private taskId2: string = "";
   private activeUserSearchindex = 1;
-  private UserSearchListLabel: UserSearchListLabelPayload[] = SEARCH_USERS_BY;
+
   private isUserAllowed: boolean = false
   private containerHeight: number = 0;
   private taskScrollableHeight: string = '100px';
@@ -697,33 +280,7 @@ export default class Tasklist extends Mixins(TaskListMixin) {
     this.userName = getUserName();
   }
 
-  timedifference(date: Date) {
-    return moment(date).fromNow();
-  }
-
-  async toggleassignee() {
-    this.loadingEditAssignee=true;
-    const reviewerList = await CamundaRest.getUsersByMemberGroups(
-      this.token,
-      this.bpmApiUrl,
-      reviewerGroup
-    );
-    if (reviewerList) {
-      this.reviewerUsersList = [];
-      reviewerList.data.forEach((user: UserPayload) => {
-        this.reviewerUsersList.push(UserListObject(user));
-      });
-      const userList = JSON.parse(JSON.stringify(this.reviewerUsersList));
-      this.userSelected = userList.find(((user: any) => user.code?.includes(this.task.assignee)));
-      this.loadingEditAssignee=false;
-      this.editAssignee = !this.editAssignee;
-    }
-  }
-
-  setSelectedUserSearchBy(searchby: string, index: number) {
-    this.selectSearchType = searchby;
-    this.activeUserSearchindex = index;
-  }
+ 
 
   async onFormSubmitCallback(actionType = "") {
     if (this.task.id !== null) {
@@ -732,55 +289,8 @@ export default class Tasklist extends Mixins(TaskListMixin) {
     }
   }
 
-  addGroup() {
-    CamundaRest.createTaskGroupByID(
-      this.token,
-      this.task.id!,
-      this.bpmApiUrl,
-      {
-        userId: null,
-        groupId: this.setGroup,
-        type: "candidate",
-      }
-    ).then(() => {
-      this.getGroupDetails();
-      this.reloadCurrentTask();
-    });
-    this.setGroup = null;
-  }
-
-  async getGroupDetails() {
-    const grouplist = await CamundaRest.getTaskGroupByID(
-      this.token,
-      this.task.id!,
-      this.bpmApiUrl
-    );
-    this.groupList = grouplist.data;
-    this.groupListItems = [];
-    this.groupListNames = undefined;
-    for (const group of grouplist.data) {
-      this.groupListItems.push(group.groupId);
-    }
-    if (this.groupListItems.length) {
-      this.groupListNames = this.groupListItems;
-    }
-  }
-
-  async deleteGroup(groupid: string) {
-    await CamundaRest.deleteTaskGroupByID(
-      this.token,
-      this.task.id!,
-      this.bpmApiUrl,
-      {
-        groupId: groupid,
-        type: "candidate",
-      }
-    ).then(async () => {
-      await this.getGroupDetails();
-      await this.reloadCurrentTask();
-    });
-  }
-
+ 
+ 
   async onBPMTaskFormSubmit(taskId: string, actionType: string) {
     let formRequestFormat: FormRequestPayload = {
       variables: {
@@ -829,7 +339,7 @@ export default class Tasklist extends Mixins(TaskListMixin) {
     this.task = taskResult.data;
     this.task.taskProcess = processResult.data.name;
     this.task.applicationId = applicationIdResult.data.applicationId.value;
-    await this.getGroupDetails();
+    // await this.getGroupDetails();
   }
 
   async getTaskFormIODetails(taskId: string) {
@@ -956,51 +466,14 @@ export default class Tasklist extends Mixins(TaskListMixin) {
     );
   }
 
-  async onClaim() {
-    this.loadingClaimAndUnclaim= true;
-    await CamundaRest.claim(
-      this.token,
-      this.task.id!,
-      {
-        userId: this.userName,
-      },
-      this.bpmApiUrl
-    );
-   
-    if (!SocketIOService.isConnected()) {
-      await this.getBPMTaskDetail(this.getFormsFlowTaskId);
-      await this.reloadLHSTaskList();
-    }
-    this.loadingClaimAndUnclaim= false;
 
+
+  async getBPMTaskandReload(){
+    await this.getBPMTaskDetail(this.getFormsFlowTaskId);
+    await this.reloadLHSTaskList();
   }
+  
 
-  async onUnClaim() {
-    this.loadingClaimAndUnclaim= true;
-    await CamundaRest.unclaim(this.token, this.task.id!, this.bpmApiUrl);
-
-    if (!SocketIOService.isConnected()) {
-      await this.getBPMTaskDetail(this.getFormsFlowTaskId);
-      await this.reloadLHSTaskList();
-    }
-    this.loadingClaimAndUnclaim= false;
-  }
-
-  async onSetassignee() {
-    await CamundaRest.setassignee(
-      this.token,
-      this.task.id!,
-      {
-        userId: this.userSelected?.code,
-      },
-      this.bpmApiUrl
-    );
-    await this.toggleassignee();
-    if (!SocketIOService.isConnected()) {
-      await this.getBPMTaskDetail(this.getFormsFlowTaskId);
-      await this.reloadLHSTaskList();
-    }
-  }
 
   async fetchFullTaskList(filterId: string, requestData: Payload) {
     const taskList = await CamundaRest.filterTaskList(
@@ -1045,106 +518,7 @@ export default class Tasklist extends Mixins(TaskListMixin) {
     }
   }
 
-  async onUserSearch(search: string, loading: any) {
-    if (search.length) {
-      loading(true);
-      this.reviewerUsersList = [];
-    }
 
-    if (this.selectSearchType === SEARCH_OPTION_TYPE.FIRST_NAME) {
-      const firstNameUserList = await CamundaRest.getUsersByFirstNameGroups(
-        this.token,
-        this.bpmApiUrl,
-        search,
-        reviewerGroup
-      );
-      this.reviewerUsersList = [];
-      firstNameUserList.data.map((user: UserPayload) => {
-        this.reviewerUsersList.push(UserListObject(user));
-      });
-      loading(false);
-    }
-
-    if (this.selectSearchType === SEARCH_OPTION_TYPE.LAST_NAME) {
-      const lastNameUserList = await CamundaRest.getUsersByLastNameGroups(
-        this.token,
-        this.bpmApiUrl,
-        search,
-        reviewerGroup
-      );
-      this.reviewerUsersList = [];
-      lastNameUserList.data.map((user: UserPayload) => {
-        this.reviewerUsersList.push(UserListObject(user));
-      });
-      loading(false);
-    }
-
-    if (this.selectSearchType === SEARCH_OPTION_TYPE.EMAIL) {
-      const emailUserList = await CamundaRest.getUsersByEmailGroups(
-        this.token,
-        this.bpmApiUrl,
-        search,
-        reviewerGroup
-      );
-      this.reviewerUsersList = [];
-      emailUserList.data.map((user: UserPayload) => {
-        this.reviewerUsersList.push(UserListObject(user));
-      });
-      loading(false);
-    }
-    loading(false);
-  }
-
-  async updateTaskDatedetails(taskId: string, task: TaskPayload) {
-    await CamundaRest.updateTasksByID(this.token, taskId, this.bpmApiUrl, task);
-    if (!SocketIOService.isConnected()) {
-      await this.reloadCurrentTask();
-    }
-  }
-
-  async updateFollowUpDate() {
-    const referenceobject = this.task;
-    try {
-      if (this.task?.followUp !== null) {
-        referenceobject.followUp = getISODateTime(this.task?.followUp);
-        await this.updateTaskDatedetails(this.task.id!, referenceobject);
-      }
-    } catch {
-      console.warn("Follow date error"); // eslint-disable-line no-console
-    }
-  }
-
-  async updateDueDate() {
-    const referenceobject = this.task;
-    try {
-      if (this.task?.due !== null) {
-        referenceobject.due = getISODateTime(this.task.due);
-        await this.updateTaskDatedetails(this.task.id!, referenceobject);
-      }
-    } catch {
-      console.warn("Due date error"); // eslint-disable-line no-console
-    }
-  }
-
-  async removeDueDate() {
-    const referenceobject = this.task;
-    try {
-      referenceobject["due"] = null;
-      await this.updateTaskDatedetails(this.task.id!, referenceobject);
-    } catch {
-      console.warn("Due date error"); // eslint-disable-line no-console
-    }
-  }
-
-  async removeFollowupDate() {
-    const referenceobject = this.task;
-    try {
-      referenceobject["followUp"] = null;
-      await this.updateTaskDatedetails(this.task.id!, referenceobject);
-    } catch {
-      console.warn("Follow up date error"); // eslint-disable-line no-console
-    }
-  }
 
   async fetchTaskDetails(taskId: string) {
     await Promise.all([
@@ -1205,7 +579,6 @@ export default class Tasklist extends Mixins(TaskListMixin) {
     this.setFormsFlowactiveIndex(NaN);
     this.taskLoading = true;
     this.$root.$on("call-fetchTaskDetails", async (para: any) => {
-      this.editAssignee = false;
       this.singleTaskLoading = true;
       this.setFormsFlowTaskId(para.selectedTaskId);
       await this.fetchTaskDetails(this.getFormsFlowTaskId);
@@ -1338,9 +711,7 @@ export default class Tasklist extends Mixins(TaskListMixin) {
 
   }
 
-  getExactDate(date: Date) {
-    return getFormattedDateAndTime(date);
-  }
+ 
 
   beforeDestroy() {
     SocketIOService.disconnect();
@@ -1363,126 +734,10 @@ export default class Tasklist extends Mixins(TaskListMixin) {
 }
 </script>
 
-<style lang="scss" scoped>
-.task-details-empty {
-  background: #fff;
-  margin-left: 4px;
-}
-.ctf-task-details-container {
-  // margin-left: 0.5rem;
-  background: #fff;
-  .task-title {
-    padding: 1rem 1.5rem;
-    border-top-left-radius: 0.5rem;
-    border-top-right-radius: 0.5rem;
-    h3 {
-      color: #fff;
-    }
-  }
-  .task-details {
-    height: 100px;
-    overflow-y: auto;
-  }
-  .task-date-picker {
-    .input-group {
-      width: calc(100% - 24px);
-      .form-control {
-        border-radius: 0.5rem;
-      }
-      i {
-        margin-left: -24px !important;
-        margin-top: 12px;
-        z-index: 3;
-      }
-    }
-  }
-  .task-icon-btn {
-    min-height: unset !important;
-    padding-top: 0;
-    padding-bottom: 0px;
-  }
-  .task-assignee {
-    flex: 0.5;
-    .select-assignee {
-      width: 100%;
-      min-width: 180px;
-      height: 40px;
-    }
-    .assignee-name {
-      white-space: pre;
-    }
-    .btn {
-      min-height: 40px;
-    }
-  }
-  .task-groups {
-    .group-name {
-      white-space: normal;
-      max-width: 250px;
-    }
-  }
-  .task-groups-modal {
-    .group-name-input {
-      margin-right: 1rem;
-    }
-    .add-group-btn {
-      min-width: 144px;
-    }
-    .added-group-chip {
-      background: #e7e7e7;
-      border-radius: 50rem;
-      padding: 0.25rem 0.75rem;
-      margin: 0.5rem 0.25rem;
-      cursor: pointer;
-      &:hover {
-        background: #ceeaf1;
-      }
-    }
-  }
-  .task-tabs {
-    .nav-link {
-      height: 40px;
-    }
-    .active {
-      background: var(--bs-primary);
-      color: #fff;
-      font-weight: 600;
-    }
-  }
-  .task-tab-content {
-    background: #fafafa;
-    padding: 1rem;
-    .form-tab-content {
-      &.disabled {
-        pointer-events: none;
-        opacity: 0.33;
-      }
-    }
-  }
-}
-.cft-bpmn-viewer-container {
-  min-height: 400px;
-  position: relative;
-}
-.cft-bpm-container {
-  height: 100%;
-  position: absolute;
-  width: 100%;
-  overflow: hidden;
-}
+ 
 
-.cft-grab-cursor {
-  cursor: move;
-  cursor: grab;
-}
-.cft-btn_zoom{
-  position: absolute;
-  bottom: 10em;
-  right:5%;
-}
-</style>
 <style>
-.highlight:not(.djs-connection) > .djs-visual > :nth-child(1) {
-  fill: rgb(56, 89, 138) !important;
-}
+.highlight:not(.djs-connection) .djs-visual > :nth-child(1) {
+    fill: rgb(56,89,138) !important;
+  }
 </style>
