+++ camunda-formio-tasklist-vue/src/components/layout/LeftSider.vue
@@ -88,9 +88,11 @@
               v-if="task.followUp"
             >Follow up in {{ timedifference(task.followUp) }}</div>
           </div>
-          <task-variable v-if="task._embedded&& task._embedded.variable" 
-          :variables="task._embedded.variable" 
-          :filterTaskVariable="selectedFilterTaskVariable"/>
+
+          <task-variable v-if="task._embedded&& task._embedded.variable"
+           :variables="task._embedded.variable"
+           :filterTaskVariable="selectedFilterTaskVariable"/>
+ 
         </div>
       </template>
       <div
